+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ## [Unreleased]
 
 ### Added
+- Added `overall_voice_casting_prompt` case to `additional_voice_casting_instructions` to add to overall casting prompt
 - Added `additional_voice_casting_instructions` field to voice_library_config to specific additional casting instructions for each provider
 - Added user-configurable voice library directories, and merge logic, so that project-level voices can be modified / added to
 - Added `voice_library_config` concept on both project and user level, and added first use case: filtering sts_ids that are included in the voice casting prompt